+++ NEWS.md
@@ -1,5 +1,10 @@
 # KMunicate (development version)
 
+### New features
+
+* New argument `.reverse` to plot 1 - survival rather than survival;
+* New argument `.ylab` to customise the label for the vertical axis.
+
 # KMunicate 0.0.2
 
 * Initial release.